split fastpath code for makereq into separate file
